requirements.txt
streamlit
gradio
plotly
pandas
google-cloud-aiplatform
google-cloud-bigquery
faiss-cpu
numpy
python-dotenv
requests
pypdf
google-adk